Move 'tune' command to 'alpha' group
