+++ cmd/beaker/alpha/parameter_space.go
@@ -1,4 +1,4 @@
-package experiment
+package alpha
 
 import (
 	"errors"
