+++ cmd/beaker/alpha/root.go
@@ -33,4 +33,5 @@ func NewAlphaCmd(
 
 	// Attach subcommands.
 	newTensorboardCmd(cmd, o, config)
+	newTuneCmd(cmd, o, config)
 }
